# api/internal/api/api.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.1-0.20260326002731-c322d1a39424 DO NOT EDIT.
package api

import (
	"github.com/labstack/echo/v5"
)

const (
	BearerAuthScopes bearerAuthContextKey = "BearerAuth.Scopes"
)

// FormError Invalid form data
type FormError struct {
	Errors map[string]string `json:"errors"`
}

// GeneralError A general API error
type GeneralError struct {
	Error string `json:"error"`
}

// AuthSpotifyResponse defines model for AuthSpotifyResponse.
type AuthSpotifyResponse struct {
	// AccessToken The access token for the my.fm user.
	AccessToken string `json:"accessToken"`
}

// InternalServerError A general API error
type InternalServerError = GeneralError

// UnauthorizedError A general API error
type UnauthorizedError = GeneralError

// UserResponse defines model for UserResponse.
type UserResponse struct {
	// Email The email of the user as per the auth provider.
	Email string `json:"email"`

	// Id The ID of the user.
	Id string `json:"id"`

	// Name The name of the user.
	Name string `json:"name"`

	// Onboarded Whether the user has finished being onboarded.
	Onboarded bool `json:"onboarded"`

	// Username The username of the user.
	Username string `json:"username"`
}

// bearerAuthContextKey is the context key for BearerAuth security scheme
type bearerAuthContextKey string

// PostAuthSpotifyJSONBody defines parameters for PostAuthSpotify.
type PostAuthSpotifyJSONBody struct {
	// Code Authorization code from spotify.
	Code string `json:"code"`
}

// PutUserProfileJSONBody defines parameters for PutUserProfile.
type PutUserProfileJSONBody struct {
	// Name The user's display name (1-50 characters).
	Name string `json:"name"`

	// Username The user's username (2-30 characters, alphanumeric/underscore/hyphen only).
	Username string `json:"username"`
}

// PostAuthSpotifyJSONRequestBody defines body for PostAuthSpotify for application/json ContentType.
type PostAuthSpotifyJSONRequestBody PostAuthSpotifyJSONBody

// PutUserProfileJSONRequestBody defines body for PutUserProfile for application/json ContentType.
type PutUserProfileJSONRequestBody PutUserProfileJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Authenticate with Spotify
	// (POST /auth/spotify)
	PostAuthSpotify(ctx *echo.Context) error
	// Get the user's info.
	// (GET /user)
	GetUser(ctx *echo.Context) error
	// Complete user onboarding.
	// (PUT /user/profile)
	PutUserProfile(ctx *echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostAuthSpotify converts echo context to params.
func (w *ServerInterfaceWrapper) PostAuthSpotify(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAuthSpotify(ctx)
	return err
}

// GetUser converts echo context to params.
func (w *ServerInterfaceWrapper) GetUser(ctx *echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetUser(ctx)
	return err
}

// PutUserProfile converts echo context to params.
func (w *ServerInterfaceWrapper) PutUserProfile(ctx *echo.Context) error {
	var err error

	ctx.Set(string(BearerAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutUserProfile(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/auth/spotify", wrapper.PostAuthSpotify)
	router.GET(baseURL+"/user", wrapper.GetUser)
	router.PUT(baseURL+"/user/profile", wrapper.PutUserProfile)

}