# api/openapi.yml
openapi: "3.0.0"
servers:
  - url: https://api.my.fm
info:
  version: 0.1.0
  title: my.fm API
  description: 'The my.fm backend service'
tags:
  - name: authentication
    description: Authentication-related operations
  - name: users
    description: User-related operations
  - name: self
    description: Operations related to the user's own user
security:
  - BearerAuth: []
paths:
  /auth/spotify:
    post:
      tags: [authentication]
      summary: Authenticate with Spotify
      description: >
        Exchanges Spotify authorization code for my.fm access token.
        If no my.fm account exists with the associated Spotify email, one is created.
      operationId: postAuthSpotify

      security: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  description: Authorization code from spotify.

      responses:
        '200':
          $ref: '#/components/responses/AuthSpotifyResponse'
        '400':
          description: Invalid authorization code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /user:
    get:
      tags: [users, self]
      summary: Get the user's info.
      description: >
        Returns information about the currently authenticated user.
      operationId: getUser

      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /user/profile:
    put:
      tags: [users, self]
      summary: Complete user onboarding.
      description: >
        Sets the user's name and username, marking them as onboarded.
        Can only be called once (cannot be called if the user is onboarded.)
      operationId: putUserProfile

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - username
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                  description: The user's display name (1-50 characters).
                username:
                  type: string
                  minLength: 2
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_-]+$'
                  description: The user's username (2-30 characters, alphanumeric/underscore/hyphen only).

      responses:
        '204':
          description: Onboarding completed successfully.
        '400':
          description: >
            Invalid request data. This can be due to:
            - User is already onboarded
            - Name is not between 1-50 characters
            - Username is not between 2-30 characters
            - Username contains invalid characters (only letters, numbers, underscores, and hyphens allowed)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FormError'
                  - $ref: '#/components/schemas/GeneralError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError' 

components:
  schemas:
    GeneralError:
      description: A general API error
      type: object
      required:
        - error
      properties:
        error:
          type: string
    FormError:
      description: Invalid form data
      type: object
      required:
        - errors
      properties:
        errors:
          type: object
          additionalProperties:
            type: string
  responses:
    InternalServerError:
      description: Unexpected internal error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
    UnauthorizedError:
      description: User is unauthorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
    UserResponse:
      description: Information about the user.
      content:
        application/json:
          schema:
            type: object
            required:
              - id
              - email
              - name
              - username
              - onboarded
            properties:
              id:
                type: string
                description: The ID of the user.
              email:
                type: string
                description: The email of the user as per the auth provider.
              name:
                type: string
                description: The name of the user.
              username:
                type: string
                description: The username of the user.
              onboarded:
                type: boolean
                description: Whether the user has finished being onboarded.
    AuthSpotifyResponse:
      description: Response containing the access token
      content:
        application/json:
          schema:
            type: object
            required:
              - accessToken
            properties:
              accessToken:
                type: string
                description: The access token for the my.fm user.
  securitySchemes:
    BearerAuth:
      description: Bearer token using a JWT
      type: http
      scheme: Bearer
      bearerFormat: JWT