# app/app/lib/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'https://api.my.fm' | (string & {});
};

/**
 * A general API error
 */
export type GeneralError = {
    error: string;
};

/**
 * Invalid form data
 */
export type FormError = {
    errors: {
        [key: string]: string;
    };
};

export type PostAuthSpotifyData = {
    body: {
        /**
         * Authorization code from spotify.
         */
        code: string;
    };
    path?: never;
    query?: never;
    url: '/auth/spotify';
};

export type PostAuthSpotifyErrors = {
    /**
     * Invalid authorization code.
     */
    400: GeneralError;
    /**
     * Unexpected internal error.
     */
    500: GeneralError;
};

export type PostAuthSpotifyError = PostAuthSpotifyErrors[keyof PostAuthSpotifyErrors];

export type PostAuthSpotifyResponses = {
    /**
     * Response containing the access token
     */
    200: {
        /**
         * The access token for the my.fm user.
         */
        accessToken: string;
    };
};

export type PostAuthSpotifyResponse = PostAuthSpotifyResponses[keyof PostAuthSpotifyResponses];

export type GetUserData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/user';
};

export type GetUserErrors = {
    /**
     * User is unauthorized.
     */
    401: GeneralError;
    /**
     * Unexpected internal error.
     */
    500: GeneralError;
};

export type GetUserError = GetUserErrors[keyof GetUserErrors];

export type GetUserResponses = {
    /**
     * Information about the user.
     */
    200: {
        /**
         * The ID of the user.
         */
        id: string;
        /**
         * The email of the user as per the auth provider.
         */
        email: string;
        /**
         * The name of the user.
         */
        name: string;
        /**
         * The username of the user.
         */
        username: string;
        /**
         * Whether the user has finished being onboarded.
         */
        onboarded: boolean;
    };
};

export type GetUserResponse = GetUserResponses[keyof GetUserResponses];

export type PutUserProfileData = {
    body: {
        /**
         * The user's display name (1-50 characters).
         */
        name: string;
        /**
         * The user's username (2-30 characters, alphanumeric/underscore/hyphen only).
         */
        username: string;
    };
    path?: never;
    query?: never;
    url: '/user/profile';
};

export type PutUserProfileErrors = {
    /**
     * Invalid request data. This can be due to: - User is already onboarded - Name is not between 1-50 characters - Username is not between 2-30 characters - Username contains invalid characters (only letters, numbers, underscores, and hyphens allowed)
     *
     */
    400: FormError | GeneralError;
    /**
     * User is unauthorized.
     */
    401: GeneralError;
    /**
     * Unexpected internal error.
     */
    500: GeneralError;
};

export type PutUserProfileError = PutUserProfileErrors[keyof PutUserProfileErrors];

export type PutUserProfileResponses = {
    /**
     * Onboarding completed successfully.
     */
    204: void;
};

export type PutUserProfileResponse = PutUserProfileResponses[keyof PutUserProfileResponses];